configure.ac: remove m4_esyscmd_s
m4_esyscmd_s is a macro only available in Autoconf 2.64, which despite
being released in 2009 is not available on a number of build targets for
the project (specifically CentOS/RHEL 6).

ca8e4c08a78 tries to remove the error produced with m4_pattern_allow,
but that just silences the sanity check.

Instead, replace m4_esyscmd_s with m4_esyscmd + manual removal of
newlines.