tighten the .gitignore patterns
I recently made a change to remove some no longer needed .gitignore rules
and generally improve the readability of that config file. Contributor
@floam noted that this config file was still too permissive and
ambiguous. This change adds additional refinements that should

a) make it easier for someone to understand why a file/directory is being
excluded, and

b) make it less likely that a mistake results in a file being inadvertently
excluded.